requirements-dev.txt
pip>=19.1
pytest>=3.5.0
setuptools-rust>=0.10.2
pytest-benchmark>=3.1.1
PyYAML >=5.3.1
